+++ validator.js
@@ -1035,10 +1035,10 @@ function isCurrency(str, options) {
 var iso8601 = /^([\+-]?\d{4}(?!\d{2}\b))((-?)((0[1-9]|1[0-2])(\3([12]\d|0[1-9]|3[01]))?|W([0-4]\d|5[0-2])(-?[1-7])?|(00[1-9]|0[1-9]\d|[12]\d{2}|3([0-5]\d|6[1-6])))([T\s]((([01]\d|2[0-3])((:?)[0-5]\d)?|24:?00)([\.,]\d+(?!:))?)?(\17[0-5]\d([\.,]\d+)?)?([zZ]|([\+-])([01]\d|2[0-3]):?([0-5]\d)?)?)?)?$/;
 /* eslint-enable max-len */
 
-var isISO8601 = function (str) {
+function isISO8601(str) {
   assertString(str);
   return iso8601.test(str);
-};
+}
 
 var notBase64 = /[^A-Z0-9+\/=]/i;