+++ index.js
@@ -258,7 +258,7 @@ var _toString2 = _interopRequireDefault(_toString);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var version = '8.0.0';
+var version = '8.1.0';
 
 var validator = {
   version: version,