+++ src/index.js
@@ -85,7 +85,7 @@ import normalizeEmail from './lib/normalizeEmail';
 
 import toString from './lib/util/toString';
 
-const version = '8.0.0';
+const version = '8.1.0';
 
 const validator = {
   version,