+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fix `require('validator/lib/isIS8601')` calls
   ([#688](https://github.com/chriso/validator.js/issues/688))
+- Added an `isLatLong()` and `isPostalCode()` validator
+  ([#684](https://github.com/chriso/validator.js/pull/684))
 - Allow comma in email display names
   ([#692](https://github.com/chriso/validator.js/pull/692))
 - Add missing string to `unescape()`